// src/models/pokemon.ts
import React from 'react'

export type Pokemon = {
    name: string;
    id: string;
    gif: string;
    large: string;
    normal: string;
    total: string;
    hp: string;
    attack: string;
    defense: string;
    sp_atk: string;
    sp_def: string;
    speed: string;
    type: Array<any>;
}